Fix verifyProperty handling of array property type and object value
When modifying the way the expectedJsonType is expanding to cover all array elements, missed verifyProperty branch for
property with type Array, but a "pseudo-array" object value.  Refactored function to reduce duplicate code, and fixed
this branch as well, so that the property mapping is not modified based on the number of elements in the value.